Move events pkg at server root
