+++ src/server/game/surviveler/aidirector.go
@@ -6,9 +6,9 @@ package surviveler
 
 import (
 	"math/rand"
+	"server/events"
 	"server/game"
 	"server/game/entities"
-	"server/game/events"
 	"server/math"
 	"time"
