+++ src/server/game/surviveler/game.go
@@ -9,8 +9,8 @@ import (
 	"os"
 	"os/signal"
 	"runtime"
+	"server/events"
 	"server/game"
-	"server/game/events"
 	"server/protocol"
 	"server/resource"
 	"sync"
