+++ src/server/game/surviveler/loop.go
@@ -6,7 +6,7 @@
 package surviveler
 
 import (
-	"server/game/events"
+	"server/events"
 	"server/messages"
 	"time"
